backend/requirements.txt
fastapi==0.120.1
uvicorn[standard]==0.38.0
python-dotenv==1.1.1

langchain==0.3.27
langchain-core==0.3.79
langchain-community==0.3.31
langchain-google-genai==2.1.12
langchain-ollama==0.3.8

tavily-python
httpx==0.28.1
psycopg[binary]==3.2.12
